refactor networks for research
I refactored the networks to make the codebase more modular. I moved
from one script for the current network I'm testing to a file for each
network that I'm testing in the experiments/ directory.

The current name of the modelbase that I'm testing is the upsampler_v0
which consists of a simple upsampling scheme for the features.